# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AddressDetail:
    properties:
      city:
        type: string
      createdAt:
        type: string
      default:
        type: boolean
      district:
        type: string
      id:
        type: string
      phone:
        type: string
      street:
        type: string
      ward:
        type: string
    type: object
...
  dto.ApiError:
    properties:
      code:
        type: string
      details:
        type: string
      stack:
        description: Hide in production
    type: object
...
  dto.ApiResponse-dto_LoginResponse:
    properties:
      data:
        $ref: '#/definitions/dto.LoginResponse'
      error:
        $ref: '#/definitions/dto.ApiError'
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.MetaInfo'
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
...
  dto.ApiResponse-dto_RefreshToken:
    properties:
      data:
        $ref: '#/definitions/dto.RefreshToken'
      error:
        $ref: '#/definitions/dto.ApiError'
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.MetaInfo'
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
  dto.ApiResponse-dto_UserDetail:
    properties:
      data:
        $ref: '#/definitions/dto.UserDetail'
      error:
        $ref: '#/definitions/dto.ApiError'
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.MetaInfo'
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
...
  dto.ErrorResp:
    properties:
      error:
        $ref: '#/definitions/dto.ApiError'
    type: object
...
  dto.LoginResponse:
    properties:
      accessToken:
        type: string
      accessTokenExpiresIn:
        type: string
      refreshToken:
        type: string
      refreshTokenExpiresAt:
        type: string
      sessionId:
        type: string
    type: object
  dto.MetaInfo:
    properties:
      method:
        type: string
      path:
        type: string
      requestId:
        type: string
      timestamp:
        type: string
    type: object
...
  dto.Pagination:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  dto.RefreshToken:
    properties:
      accessToken:
        type: string
      accessTokenExpiresAt:
        type: string
    type: object
  dto.UserDetail:
    properties:
      addresses:
        items:
          $ref: '#/definitions/dto.AddressDetail'
        type: array
      avatarId:
        type: string
      avatarUrl:
        type: string
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastLoginAt:
        type: string
      lastName:
        type: string
      locked:
        type: boolean
      passwordChangedAt:
        type: string
      phone:
        type: string
      roleCode:
        type: string
      roleId:
        type: string
      updatedAt:
        type: string
      username:
        type: string
      verifiedEmail:
        type: boolean
      verifiedPhone:
        type: boolean
    type: object
...
  models.CreateAddress:
    properties:
      city:
        type: string
      district:
        type: string
      isDefault:
        type: boolean
      phone:
        maxLength: 15
        minLength: 10
        type: string
      street:
        type: string
      ward:
        maxLength: 100
        type: string
    required:
    - city
    - district
    - phone
    - street
    type: object
...
  models.LoginModel:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 32
        minLength: 6
        type: string
      username:
        maxLength: 32
        type: string
    required:
    - password
    type: object
...
  models.RegisterModel:
    properties:
      address:
        $ref: '#/definitions/models.CreateAddress'
      email:
        maxLength: 255
        minLength: 6
        type: string
      firstName:
        maxLength: 32
        minLength: 3
        type: string
      lastName:
        maxLength: 32
        minLength: 3
        type: string
      password:
        maxLength: 32
        minLength: 6
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - address
    - email
    - password
    - phone
    - username
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: This is a sample server for a simple e-commerce API.
  title: E-Commerce API
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login to the system
      parameters:
      - description: User info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.LoginModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-dto_LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResp'
      summary: Login to the system
      tags:
      - users
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Refresh token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-dto_RefreshToken'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResp'
      summary: Refresh token
      tags:
      - users
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.RegisterModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-dto_UserDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResp'
      summary: Create a new user
      tags:
      - users
...
swagger: "2.0"